+++ VERenderer/Internal/VulkanFramebuffer.cpp
@@ -9,7 +9,7 @@ VulkanFramebuffer::VulkanFramebuffer(VulkanDevice * device, uint32_t width, uint
     VkFramebufferCreateInfo framebufferInfo = {};
     framebufferInfo.sType = VK_STRUCTURE_TYPE_FRAMEBUFFER_CREATE_INFO;
     framebufferInfo.renderPass = renderPass->getHandle();
-    framebufferInfo.attachmentCount = attachments.size();
+    framebufferInfo.attachmentCount = static_cast<uint32_t>(attachments.size());
     framebufferInfo.pAttachments = attachments.data();
     framebufferInfo.width = width;
     framebufferInfo.height = height;
