+++ VERenderer/VulkanSwapChainOutput.h
@@ -9,7 +9,7 @@ public:
     ~VulkanSwapChainOutput();
     void beginDrawing();
     void endDrawing();
-    void drawMesh(Object3dInfo *info, size_t instances);
+    void drawMesh(Object3dInfo *info, uint32_t instances);
     void submit(std::vector<VkSemaphore> waitSemaphores);
 private:
     VulkanDevice * device;
